Stop updating state file on SIGTERM
to avoid collision against a new process.
